# src/main/resources/swagger.yml
swagger: "2.0"
info:
  title: Poll System
  contact:
    name: NTT Data
    url: https://it.nttdata.com/
  version: 1.0.0
basePath: /rest/api/v0
paths:
  /login:
    post:
      summary:  API per recuperare token di auth
      operationId: postLogin
      produces:
        - application/json
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/LoginRequest"
      responses:
        "200":
          description: "Login eseguito con successo"
          schema:
            $ref: '#/definitions/LoginResponse'
        "401":
          description: "Login fallito"
  /registration:
    post:
      summary:  API per creare un nuovo user
      operationId: registration
      produces:
        - application/json
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/RegistrationRequest"
      responses:
        "201":
          description: "utente creato con successo"
        "400":
          description: "Impossibile creare l'utente"
          schema:
            $ref: '#/definitions/ValidationErrorResponse'
...
definitions:
  RegistrationRequest:
    type: object
    required:
      - username
      - password
      - email
    properties:
      username:
        type: string
        description: univoco all interno della piattaforma
      password:
        type: string
        minLength: 8
        description: lunga almeno 8 caratteri e contenente almeno un numero, una lettera minuscola e una maiuscola
      email:
        type: string
        description: deve essere univoca nella piattaforma e deve avere il formato di una mail
  LoginRequest:
    type: object
    required:
      - username
      - password
    properties:
      username:
        type: string
      password:
        type: string
  LoginResponse:
    type: object
    required:
      - token
      - expiresAt
    properties:
      token:
        type: string
        description: Codice Fattura univoco
      expiresAt:
        type: string
        format: date-time
  ValidationErrorResponse:
    type: object
    properties:
      timestamp:
        type: string
        format: date-time
      status:
        type: integer
      message:
        type: string
      path:
        type: string
...